spike/webview-notebook/requirements.txt
pywebview>=5.0,<6
